refactor: organized src/eval
