+++ src/eval/mod.rs
@@ -1,0 +1,6 @@
+mod eval;
+mod expression;
+mod literal;
+mod statement;
+
+pub use eval::Evaluator;
